Revert "feat(*): extract interfaces into separate package"
This reverts commit bebd0d1d9af23a406906406b5a06fd80c45aa1f1.
