+++ consumers/stdout.go
@@ -6,7 +6,6 @@ import (
 
 	log "github.com/Sirupsen/logrus"
 
-	"github.com/zalando-incubator/mate/interfaces"
 	"github.com/zalando-incubator/mate/pkg"
 )
 
@@ -14,7 +13,7 @@ import (
 // or new file writrer better
 type stdoutConsumer struct{}
 
-func NewStdout() (interfaces.Consumer, error) {
+func NewStdout() (Consumer, error) {
 	return &stdoutConsumer{}, nil
 }
