+++ consumers/synced.go
@@ -3,20 +3,19 @@ package consumers
 import (
 	"sync"
 
-	"github.com/zalando-incubator/mate/interfaces"
 	"github.com/zalando-incubator/mate/pkg"
 )
 
 type SyncedConsumer struct {
 	sync.Mutex
-	interfaces.Consumer
+	Consumer
 }
 
 // NewSynced provides a consumer that can execute only
 // one operation at a time, and blocks
 // concurrent operations until the current one
 // finishes.
-func NewSynced(name string) (interfaces.Consumer, error) {
+func NewSynced(name string) (Consumer, error) {
 	consumer, err := New(name)
 	if err != nil {
 		return nil, err
