+++ producers/producers.go
@@ -2,11 +2,17 @@ package producers
 
 import (
 	"fmt"
+	"sync"
 
-	"github.com/zalando-incubator/mate/interfaces"
+	"github.com/zalando-incubator/mate/pkg"
 )
 
-func New(name string) (interfaces.Producer, error) {
+type Producer interface {
+	Endpoints() ([]*pkg.Endpoint, error)
+	Monitor(chan *pkg.Endpoint, chan error, chan struct{}, *sync.WaitGroup)
+}
+
+func New(name string) (Producer, error) {
 	switch name {
 	case "kubernetes":
 		return NewKubernetes()
